+++ hc/api/tests/test_notify.py
@@ -2,7 +2,7 @@
 
 from datetime import timedelta as td
 import json
-from unittest.mock import patch, Mock
+from unittest.mock import patch
 
 from django.core import mail
 from django.utils.timezone import now
@@ -491,33 +491,6 @@ class NotifyTestCase(BaseTestCase):
         self.channel.notify(self.check)
         n = Notification.objects.first()
         self.assertEqual(n.error, 'Received status code 403 with a message: "Nice try"')
-
-    @patch("hc.api.transports.requests.request")
-    def test_pushover(self, mock_post):
-        self._setup_data("po", "123|0")
-        mock_post.return_value.status_code = 200
-
-        self.channel.notify(self.check)
-        assert Notification.objects.count() == 1
-
-        args, kwargs = mock_post.call_args
-        payload = kwargs["data"]
-        self.assertIn("DOWN", payload["title"])
-
-    @patch("hc.api.transports.requests.request")
-    def test_pushover_up_priority(self, mock_post):
-        self._setup_data("po", "123|0|2", status="up")
-        mock_post.return_value.status_code = 200
-
-        self.channel.notify(self.check)
-        assert Notification.objects.count() == 1
-
-        args, kwargs = mock_post.call_args
-        payload = kwargs["data"]
-        self.assertIn("UP", payload["title"])
-        self.assertEqual(payload["priority"], 2)
-        self.assertIn("retry", payload)
-        self.assertIn("expire", payload)
 
     @patch("hc.api.transports.requests.request")
     def test_victorops(self, mock_post):